Also extract fakeIdp to testenv package to clean up app_test a bit and make it easier to have more _test files for such tests in the future.
